Moved option parser code to the library and added unit tests for it
